Polish
This commit tries to move a bit further. In particular:

* CouchbaseProperties has been revisited to minimize the changes
compared to master. Additional metadata should be added later
to indicate that properties got removed
* Bucket name has moved to CouchbaseDataProperties as only the latter is
using it now. There is no default value anymore and a value must be
provided for the auto-configuration to kick in (not done yet)
* The CouchbaseConfigurer infrastructure was restored
* Fix for health indicators
* Fix for cache

The data test are broken due to the lack of separation between the
infrastructure and the abstract config that creates a cluster for you